2.10
Added Password Hashing, Validation, and CORS access control

+++ index.js
@@ -1,5 +1,6 @@
 //Import Node modules
 const express = require('express');
+const { check, validationResult } = require('express-validator');
 const bodyParser = require('body-parser');
 
 const morgan = require('morgan');
@@ -24,6 +25,10 @@ mongoose.connect('mongodb://127.0.0.1:27017/cfDB', {
 //Body Parser
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
+
+//Cors (Cross-origin Resource Sharing)
+const cors = require('cors');
+app.use(cors());
 
 //Import auth.js
 let auth = require('./auth')(app);
@@ -90,20 +95,32 @@ app.get('/directors/:Name', passport.authenticate('jwt', { session: false }), (r
 });
 
 //POST New User
-app.post('/users', passport.authenticate('jwt', { session: false }), (req, res) => {
+app.post('/users', passport.authenticate('jwt', { session: false }),
+[
+  check('Username', 'Username is required').isLength({min: 5}),
+  check('Username', 'Username contains non alphanumeric characters - not allowed.').isAlphanumeric(),
+  check('Password', 'Password is required').not().isEmpty(),
+  check('Email', 'Email does not appear to be valid').isEmail()
+], (req, res) => {
+  let errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }
+  
+  let hashedPassword = Users.hashPassword(req.body.Password);
   Users.findOne({ Username: req.body.Username })
   .then((user) => {
     if(user) {
       return res.status(400).send(req.body.Username + " already exists!");
     } else {
-      Users.create( {
+      Users.create({
         Username: req.body.Username,
-        Password: req.body.Password,
+        Password: hashedPassword,
         Email: req.body.Email,
         Birthday: req.body.Birthday,
-      }).then((user) => {
-        res.status(201).json(user);
-      }).catch((err) => {
+      })
+      .then((user) => { res.status(201).json(user);})
+      .catch((err) => {
         console.error(err);
         res.status(500).send("Error: " + err);
       });
@@ -112,11 +129,23 @@ app.post('/users', passport.authenticate('jwt', { session: false }), (req, res) 
 });
 
 //PUT Update information of user
-app.put('/users/:Username', passport.authenticate('jwt', { session: false }), (req, res) => {
+app.put('/users/:Username', passport.authenticate('jwt', { session: false }),
+[
+  check('Username', 'Username is required').isLength({min: 5}),
+  check('Username', 'Username contains non alphanumeric characters - not allowed.').isAlphanumeric(),
+  check('Password', 'Password is required').not().isEmpty(),
+  check('Email', 'Email does not appear to be valid').isEmail()
+], (req, res) => {
+  let errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }  
+  
+  let hashedPassword = Users.hashPassword(req.body.Password)
   Users.findOneAndUpdate({ Username: req.params.Username }, { $set:      
       {
         Username: req.body.Username,
-        Password: req.body.Password,
+        Password: hashedPassword,
         Email: req.body.Email,
         Birthday: req.body.Birthday,
       }
@@ -130,7 +159,16 @@ app.put('/users/:Username', passport.authenticate('jwt', { session: false }), (r
 });
 
 //Post Add movie to user's favorites
-app.post('/users/:Username/favorites/:MovieID', passport.authenticate('jwt', { session: false }), (req, res) => {
+app.post('/users/:Username/favorites/:MovieID', passport.authenticate('jwt', { session: false }),
+[
+  check('Username', 'Username is required').isLength({min: 5}),
+  check('Username', 'Username contains non alphanumeric characters - not allowed.').isAlphanumeric(),
+], (req, res) => {
+  let errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }  
+  
   Users.findOneAndUpdate({ Username: req.params.Username }, { $push:      
     {
       Favorites: req.params.MovieID
@@ -142,11 +180,20 @@ app.post('/users/:Username/favorites/:MovieID', passport.authenticate('jwt', { s
     } else {
       res.status(500).send("Error occurred updating user.");
     }
-  })
+  })  
 });
 
 //DELETE Remove movie from favorites
-app.delete('/users/:Username/favorites/:MovieID', passport.authenticate('jwt', { session: false }), (req, res) => {
+app.delete('/users/:Username/favorites/:MovieID', passport.authenticate('jwt', { session: false }),
+[
+  check('Username', 'Username is required').isLength({min: 5}),
+  check('Username', 'Username contains non alphanumeric characters - not allowed.').isAlphanumeric()
+], (req, res) => {
+  let errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }  
+  
   Users.findOneAndUpdate({ Username: req.params.Username }, { $pull: 
     { 
       Favorites: req.params.MovieID 
@@ -162,7 +209,16 @@ app.delete('/users/:Username/favorites/:MovieID', passport.authenticate('jwt', {
 });
 
 //DELETE Deregister user by username
-app.delete('/users/:Username', passport.authenticate('jwt', { session: false }), (req, res) => {
+app.delete('/users/:Username', passport.authenticate('jwt', { session: false }), 
+[
+  check('Username', 'Username is required').isLength({min: 5}),
+  check('Username', 'Username contains non alphanumeric characters - not allowed.').isAlphanumeric()
+], (req, res) => {
+  let errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }  
+  
   Users.findOneAndRemove({ Username: req.params.Username }).then((user) => {
     if (user) {
       res.status(201).send('User ' + req.params.Username + ' has been deleted.');
